Integrated the post-process clean up into the classification_vis script
