+++ .gitignore
@@ -5,4 +5,10 @@ ConeFlareRe45.keras
 ConeFlareRe33.keras
 ConeFlareRe45_normal.keras
 ConeFlareRe33_normal.keras
-
+Confidence_class_Re45.npy
+Confidence_class_Run38.npy
+Confidence_class.npy
+training_data_LangleyRun38.txt
+True_class_Re45.npy
+True_class_Run38.npy
+run38_trained_classifier_RandomSampled.keras
